pyproject.toml
[project]
name = "sliders-genie"
version = "0.1.0"
description = "SLIDERS: scalable question answering over long document sets via schema-guided extraction and SQL-based reconciliation."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Harshit Joshi", email = "hj@cs.stanford.edu" },
    { name = "Priyank Shethia" },
    { name = "Jadelynn Dao" },
    { name = "Monica S. Lam" },
]
maintainers = [
    { name = "Harshit Joshi", email = "hj@cs.stanford.edu" },
]
keywords = [
    "llm",
    "rag",
    "question-answering",
    "information-extraction",
    "long-context",
    "sql",
    "document-ai",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
]
dependencies = [
    "docling>=2.0",
    "duckdb>=1.2.2",
    "jinja2>=3.1",
    "langchain>=0.3.3",
    "langchain-community>=0.3.2",
    "langchain-openai>=0.2.2",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.3.33",
    "loguru>=0.7.2",
    "marko>=2.1.3",
    "markdown>=3.8",
    "matplotlib>=3.9.2",
    "numpy>=1.26",
    "openai>=1.30",
    "opencv-python-headless>=4.8",
    "overrides>=7.7",
    "pandas>=2.2",
    "pydantic>=2.7",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "redis>=5.2.0",
    "retry>=0.9.2",
    "rich>=13.9.4",
    "tenacity>=8.3",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.3.4",
    "ruff>=0.6",
]

[project.scripts]
sliders = "sliders.cli:main"

[project.urls]
Homepage = "https://sliders.genie.stanford.edu"
Repository = "https://github.com/stanford-oval/sliders-public"
"Lab" = "https://oval.cs.stanford.edu"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sliders"]
# Non-Python assets under `sliders/` (prompts, YAML config, taxonomy JSON) are
# picked up automatically because they live inside the package directory.

[tool.hatch.build.targets.sdist]
include = [
    "sliders",
    "README.md",
    "LICENSE",
    "pyproject.toml",
    "sample_docs",
    "examples",
    ".example.env",
]
exclude = [
    "**/__pycache__",
    "**/*.pyc",
]

# Force the headless OpenCV wheel so Docling works in server/container environments
# that don't ship libGL (both packages provide the same `cv2` module).
[tool.uv]
override-dependencies = [
    "opencv-python ; sys_platform == 'never'",
]

[tool.ruff]
line-length = 120
indent-width = 4
